BugFix: Inject permissions when joining as OP
Fix the following issue:
If a player has the "permissionsex.disabled" permission, PermissionsEx does not inject permissions for this player. Due to a change in Bukkit, being OP gives a player all permissions, resulting in PermissionsEx not injecting permissions for this player and leading to unexpected behaviour as soon as the given player de-OPs.